Revert "feat: centralize segmented control glass helper"
